# docs/02-openapi-spec.yaml
openapi: 3.0.3
info:
  title: TikTok Shop AI Video System API
  version: 1.0.0
  description: |
    Java Spring Boot 主业务后端 API。
    前端只调用本 API。
    Python AI 服务和 Render Worker 通过 callback 接口回调本 API。
    V1 契约版本为 1.0.0。AI 回调、RenderManifest 和数据库记录必须保存对应版本号。
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/video-tasks/{taskId}/assets:
    post:
      tags: [VideoTasks]
      summary: Add asset to task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "200":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
    get:
      tags: [VideoTasks]
      summary: List task assets
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
  /api/video-tasks/{taskId}/assets/{assetId}/role:
    patch:
      tags: [VideoTasks]
      summary: Update asset role
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: assetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
  /api/video-tasks/{taskId}/assets/{assetId}:
    delete:
      tags: [VideoTasks]
      summary: Delete task asset
      description: Delete an uploaded task asset while the task is still in asset upload or asset confirmation stage.
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: assetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
  /api/video-tasks/{taskId}/assets/generate-image:
    post:
      tags: [VideoTasks]
      summary: Generate or edit a task-level product image
      description: Generate one AI image candidate from existing task assets and user instruction. The generated image is saved as an unconfirmed task asset with source=ai_generated.
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAssetImageRequest"
      responses:
        "200":
          description: Generated asset saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
  /api/video-tasks/{taskId}/assets/{assetId}/regenerate-image:
    post:
      tags: [VideoTasks]
      summary: Regenerate an AI-generated task image with user feedback
      description: Create a new unconfirmed AI image candidate from a previous generated asset, preserving regeneration context in task asset metadata.
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: assetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateAssetImageRequest"
      responses:
        "200":
          description: Regenerated asset saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssetListResponse"
  /api/video-tasks/{taskId}/assets/confirm:
    post:
      tags: [VideoTasks]
      summary: Confirm assets
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmAssetsRequest"
      responses:
        "200":
          description: Assets confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
...
  /api/video-tasks/{taskId}/video-clips:
    get:
      tags: [VideoTasks]
      summary: List video clips
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Video clip list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoClipListResponse"
  /api/video-tasks/{taskId}/video-clips/{clipId}/confirm:
    post:
      tags: [VideoTasks]
      summary: Confirm a video clip
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: clipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmVideoClipRequest"
      responses:
        "200":
          description: Clip confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
  /api/video-tasks/{taskId}/video-clips/{clipId}/reject:
    post:
      tags: [VideoTasks]
      summary: Reject a video clip
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: clipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmVideoClipRequest"
      responses:
        "200":
          description: Clip rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTaskStatusResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    VideoTaskStatus:
      type: string
      enum:
        - draft
        - asset_uploading
        - asset_analyzing
        - waiting_asset_confirmation
        - reference_analyzing
        - plan_generating
        - analyzing
        - analysis_completed
        - plan_generated
        - waiting_plan_selection
        - storyboard_generating
        - script_generating
        - script_generated
        - material_generating
        - material_generated
        - rendering
        - checking
        - completed
        - failed
        - exported
        - waiting_storyboard_confirmation
        - keyframe_configuring
        - image_generating
        - waiting_image_confirmation
        - video_clip_generating
        - waiting_video_clip_confirmation
        - waiting_final_review
        - repairing
        - cancelled
...
    VideoTaskStatusResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/VideoTaskStatus"
            progress:
              type: integer
...
    TaskAsset:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
          nullable: true
        shotId:
          type: string
          format: uuid
          nullable: true
        assetKind:
          type: string
          enum: [image, video, audio]
        url:
          type: string
        assetRole:
          type: string
          enum:
            - product_front
            - product_back
            - product_detail
            - model_reference
            - scene_reference
            - outfit_reference
            - reference_video
            - user_keyframe
            - generated_result
            - ai_keyframe
            - image_variant
            - video_clip
            - final_video
            - cover_image
        source:
          type: string
          enum: [user_upload, ai_generated, external_url]
        fileName:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        confirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VideoClip:
      type: object
      properties:
        clipId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        shotId:
          type: string
          format: uuid
        keyframeId:
          type: string
          format: uuid
          nullable: true
        shotNo:
          type: integer
        source:
          type: string
          enum: [user_upload, ai_generated]
        url:
          type: string
          nullable: true
        prompt:
          type: string
          nullable: true
        negativePrompt:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        modelName:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, generating, generated, uploaded, confirmed, rejected, failed]
        duration:
          type: integer
        version:
          type: integer
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VideoClipListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            clips:
              type: array
              items:
                $ref: "#/components/schemas/VideoClip"
    ConfirmVideoClipRequest:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
        feedback:
          type: string
...
    TaskAssetListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
            assets:
              type: array
              items:
                $ref: "#/components/schemas/TaskAsset"
    UpdateAssetRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum:
            - product_front
            - product_back
            - product_detail
            - model_reference
            - scene_reference
            - outfit_reference
            - reference_video
            - user_keyframe
            - generated_result
            - ai_keyframe
            - image_variant
            - video_clip
            - final_video
            - cover_image
        shotId:
          type: string
          format: uuid
          nullable: true
    ConfirmAssetsRequest:
      type: object
      properties:
        assetIds:
          type: array
          items:
            type: string
            format: uuid
    GenerateAssetImageRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        sourceAssetIds:
          type: array
          items:
            type: string
            format: uuid
        assetRole:
          type: string
          enum: [generated_result, image_variant]
          default: image_variant
    RegenerateAssetImageRequest:
      type: object
      required: [feedback]
      properties:
        feedback:
          type: string
          minLength: 1
...
    CreateAssetRequest:
      type: object
      required: [assetKind, assetRole, url, source]
      properties:
        assetKind:
          type: string
          enum: [image, video, audio]
        assetRole:
          type: string
        url:
          type: string
        source:
          type: string
          enum: [user_upload, ai_generated, external_url]
        fileName:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
...